// src/types/global.ts
export enum AvailableSidebars {
    ApplyForm = 'ApplyForm',
    TermOfUse = 'TermOfUse',
    FilterPros = 'FilterPros'
}
export enum SocialNetworks {
    facebook = 'facebook',
    instagram = 'instagram',
    linkedin = 'linkedin',
    twitter = 'twitter',
    tiktok = 'tiktok',
    youtube = 'youtube',
    pinterest = 'pinterest',
    general = 'general'
}